+++ src/lib/events/eventLinkTransfer.js
@@ -12,6 +12,12 @@ const fields = [
     description: l10n("FIELD_LINK_TRANSFER_SEND_VARIABLE_DESC"),
     type: "variable",
     defaultValue: "LAST_VARIABLE",
+    conditions: [
+      {
+        key: "size",
+        eq: 1,
+      },
+    ],
   },
   {
     key: "receiveVariable",
@@ -19,6 +25,40 @@ const fields = [
     description: l10n("FIELD_LINK_TRANSFER_RECEIVE_VARIABLE_DESC"),
     type: "variable",
     defaultValue: "LAST_VARIABLE",
+    conditions: [
+      {
+        key: "size",
+        eq: 1,
+      },
+    ],
+  },
+  {
+    key: "sendVariableArray",
+    label: l10n("FIELD_SEND_VARIABLE"),
+    description: l10n("FIELD_LINK_TRANSFER_SEND_VARIABLE_DESC"),
+    type: "variable",
+    variableType: "arrayReference",
+    defaultValue: "LAST_VARIABLE",
+    conditions: [
+      {
+        key: "size",
+        gt: 1,
+      },
+    ],
+  },
+  {
+    key: "receiveVariableArray",
+    label: l10n("FIELD_RECEIVE_VARIABLE"),
+    description: l10n("FIELD_LINK_TRANSFER_RECEIVE_VARIABLE_DESC"),
+    type: "variable",
+    variableType: "arrayReference",
+    defaultValue: "LAST_VARIABLE",
+    conditions: [
+      {
+        key: "size",
+        gt: 1,
+      },
+    ],
   },
   {
     key: "size",
@@ -26,12 +66,21 @@ const fields = [
     description: l10n("FIELD_LINK_TRANSFER_PACKET_SIZE_DESC"),
     type: "number",
     defaultValue: 1,
+    min: 1,
   },
 ];
 
 const compile = (input, helpers) => {
   const { linkTransfer } = helpers;
-  linkTransfer(input.sendVariable, input.receiveVariable, input.size);
+  if (input.size === 1) {
+    linkTransfer(input.sendVariable, input.receiveVariable, input.size);
+  } else {
+    linkTransfer(
+      input.sendVariableArray,
+      input.receiveVariableArray,
+      input.size,
+    );
+  }
 };
 
 module.exports = {